+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `BaseImage.size`
 - Support for terminal size-relative padding ([#91]).
 - `ANIM` render method to the `iterm2` render style ([#92]).
+- `term_image.exceptions.RenderError` ([#94]).
 
 ### Changed
 - `UrwidImage.clear_all()` -> `UrwidImageScreen.clear_images()` ([08f4e4d]).
@@ -42,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Changed default values to `0` and `-2` respectively.
 - Swapped `N` for `A` in the *method* field of the `iterm2` style-speific render format specification ([#92]).
 - `StyleError` is now raised instead of style-specific exceptions ([#93]).
+- **(BREAKING!)** `term_image.exceptions.RenderError` is now raised for errors that occur during rendering ([#94]).
 
 ### Removed
 - Image scaling ([#88]).
@@ -66,6 +68,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#91]: https://github.com/AnonymouX47/term-image/pull/91
 [#92]: https://github.com/AnonymouX47/term-image/pull/92
 [#93]: https://github.com/AnonymouX47/term-image/pull/93
+[#94]: https://github.com/AnonymouX47/term-image/pull/94
 [08f4e4d]: https://github.com/AnonymouX47/term-image/commit/08f4e4d1514313bbd4278dadde46d21d0b11ed1f
 [fa47742]: https://github.com/AnonymouX47/term-image/commit/fa477424c83474256d4972c4b2cdd4a765bc1cda
 [ed3baa3]: https://github.com/AnonymouX47/term-image/commit/ed3baa38d7621720c007f4662f89d7abadd76ec3